move BaseExample into `docs` package
