+++ packages/core/examples/collapseExample.tsx
@@ -8,7 +8,7 @@
 import * as React from "react";
 
 import { Button, Collapse } from "@blueprintjs/core";
-import BaseExample from "./common/baseExample";
+import BaseExample from "@blueprintjs/docs/dist/components/baseExample";
 
 export class CollapseExample extends BaseExample<{ isOpen: boolean }> {
     public state = {
