+++ packages/core/examples/dropdownMenuExample.tsx
@@ -8,7 +8,7 @@
 import * as React from "react";
 
 import { Menu, MenuDivider, MenuItem, Popover, Position } from "@blueprintjs/core";
-import BaseExample from "./common/baseExample";
+import BaseExample from "@blueprintjs/docs/dist/components/baseExample";
 
 export class DropdownMenuExample extends BaseExample<{}> {
     protected renderExample() {
