+++ packages/core/examples/focusExample.tsx
@@ -8,7 +8,7 @@
 import * as React from "react";
 
 import { Button, FocusStyleManager, InputGroup, Switch } from "@blueprintjs/core";
-import BaseExample, { handleBooleanChange } from "./common/baseExample";
+import BaseExample, { handleBooleanChange } from "@blueprintjs/docs/dist/components/baseExample";
 
 export interface IFocusExampleState {
     isFocusActive?: boolean;
