+++ packages/core/examples/hotkeyPiano.tsx
@@ -11,7 +11,7 @@ import * as classNames from "classnames";
 import * as React from "react";
 
 import { Hotkey, Hotkeys, HotkeysTarget } from "@blueprintjs/core";
-import BaseExample from "./common/baseExample";
+import BaseExample from "@blueprintjs/docs/dist/components/baseExample";
 
 class Oscillator {
     public oscillator: OscillatorNode;
