+++ packages/core/examples/hotkeyTester.tsx
@@ -8,7 +8,7 @@
 import * as React from "react";
 
 import { getKeyComboString, KeyCombo } from "@blueprintjs/core";
-import BaseExample from "./common/baseExample";
+import BaseExample from "@blueprintjs/docs/dist/components/baseExample";
 
 export interface IHotkeyTesterState {
     combo: string;
