+++ packages/core/examples/popoverExample.tsx
@@ -21,7 +21,7 @@ import {
     Slider,
     Switch,
 } from "@blueprintjs/core";
-import BaseExample, { handleBooleanChange, handleNumberChange } from "./common/baseExample";
+import BaseExample, { handleBooleanChange, handleNumberChange } from "@blueprintjs/docs/dist/components/baseExample";
 
 const INTERACTION_KINDS = [
     { label: "Click", value: PopoverInteractionKind.CLICK.toString() },
