+++ packages/core/examples/progressExample.tsx
@@ -8,7 +8,7 @@
 import * as React from "react";
 
 import { Intent, ProgressBar, Slider, Switch } from "@blueprintjs/core";
-import BaseExample, { handleBooleanChange, handleNumberChange } from "./common/baseExample";
+import BaseExample, { handleBooleanChange, handleNumberChange } from "@blueprintjs/docs/dist/components/baseExample";
 import { IntentSelect } from "./common/intentSelect";
 
 export interface IProgressExampleState {
