+++ packages/core/examples/rangeSliderExample.tsx
@@ -8,7 +8,7 @@
 import * as React from "react";
 
 import { NumberRange, RangeSlider } from "@blueprintjs/core";
-import BaseExample from "./common/baseExample";
+import BaseExample from "@blueprintjs/docs/dist/components/baseExample";
 
 export interface IRangeSliderExampleState {
     range?: NumberRange;
