+++ packages/core/examples/spinnerExample.tsx
@@ -8,7 +8,7 @@
 import * as React from "react";
 
 import { Classes, Spinner } from "@blueprintjs/core";
-import { handleStringChange } from "./common/baseExample";
+import { handleStringChange } from "@blueprintjs/docs/dist/components/baseExample";
 import { ProgressExample } from "./progressExample";
 
 const SIZES = [
