+++ packages/core/examples/tabs2Example.tsx
@@ -9,7 +9,7 @@ import * as classNames from "classnames";
 import * as React from "react";
 
 import { Classes, Switch, Tab2, Tabs2 } from "@blueprintjs/core";
-import BaseExample, { handleBooleanChange } from "./common/baseExample";
+import BaseExample, { handleBooleanChange } from "@blueprintjs/docs/dist/components/baseExample";
 
 export interface ITabs2ExampleState {
     activeTabId?: string;
