+++ packages/core/examples/tagExample.tsx
@@ -8,7 +8,7 @@
 import * as React from "react";
 
 import { Classes, Intent, Tag } from "@blueprintjs/core";
-import BaseExample from "./common/baseExample";
+import BaseExample from "@blueprintjs/docs/dist/components/baseExample";
 
 export class TagExample extends BaseExample<{ showTag?: boolean }> {
     public state = {
