+++ packages/core/examples/tooltipExample.tsx
@@ -8,7 +8,7 @@
 import * as React from "react";
 
 import { Intent, Popover, Position, Switch, Tooltip } from "@blueprintjs/core";
-import BaseExample from "./common/baseExample";
+import BaseExample from "@blueprintjs/docs/dist/components/baseExample";
 
 export class TooltipExample extends BaseExample<{ isOpen: boolean }> {
     public state = {
