+++ packages/datetime/examples/tsconfig.json
@@ -1,6 +1,7 @@
 {
     "version": "2.0.0",
     "compilerOptions": {
+        "baseUrl": ".",
         "declaration": true,
         "experimentalDecorators": true,
         "jsx": "react",
@@ -14,6 +15,9 @@
         "removeComments": false,
         "sourceMap": false,
         "stripInternal": true,
-        "target": "es5"
+        "target": "es5",
+        "paths": {
+            "@blueprintjs/docs/*": ["../../docs/*"]
+        }
     }
 }
