+++ packages/docs/src/components/baseExample.tsx
