+++ packages/table/examples/cellLoadingExample.tsx
@@ -8,7 +8,7 @@
 import * as React from "react";
 
 import { RadioGroup } from "@blueprintjs/core";
-import BaseExample, { handleStringChange } from "@blueprintjs/core/examples/common/baseExample";
+import BaseExample, { handleStringChange } from "@blueprintjs/docs/dist/components/baseExample";
 
 import { Cell, Column, ColumnHeaderCell, RowHeaderCell, Table } from "../src";
